third_party/WebKit/LayoutTests/fast/dom/HTMLImageElement/image-async-loading-data-uris.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
    test(function() {
        var img = new Image();
        img.src = "data:image/gif;base64,R0lGODlhAQABAIAAAAUEBAAAACwAAAAAAQABAAACAkQBADs=";
        assert_equals(img.width, 0);
    }, "Test that a data URI resource doesn't get loaded synchronously");
</script>
